requirements.txt
albumentations==1.3.0
krippendorff==0.6.0
matplotlib==3.5.1
numpy==1.21.5
opencv_python_headless==4.7.0.72
pandas==1.4.2
Pillow==9.5.0
scikit_learn==1.2.2
torch==1.13.1+cu116
torchmetrics==0.11.4
torchsummary==1.5.1
torchvision==0.14.1+cu116
torchviz==0.0.2
tqdm==4.64.0
